Major building infrastructure update
Now buildings and add-ons work as intended. Mostly. This has been a major update. Now focus is on getting the main panel functional and isolate the culture panel to its own tab. After that adding a new vassal button and adding officials is the focus. The next step after that is making the units tab, although that is much later!