reworking buildings and hexpane a bit
redoing the buildings frame, is doing some methods in conjunction with the construction of the initial frame. Have constructed the buildingspanel as well as addbuildingrow and removebuildingrow. continue work on fortificationspanel and addfortificationrow as well as removefortificationsrow. Currently creating a recursive loop upon activating the building button. need to add some outputs to pinpoint the error.